cmake: convert FindNetCDF.cmake for imported targets
A recent blog post can be found at [1].
FindNetCDF.cmake has been downloaded from [2].

This allows to add the imported target NetCDF::NetCDF_Fortran to
the target_link_libraries of fesom. The libraries (.so/.a) and the
include directories are automatically added to the target, the
target_include_directory for netcdf are implied.

[1] https://www.scivision.dev/cmake-find-netcdf4/
[2] https://github.com/geospace-code/nc4fortran/blob/master/cmake/Modules/FindNetCDF.cmake