editorconfig: initial config file
This commit introduce the file .editorconfig which is a configuration
for EditorConfig [1]:
" EditorConfig helps maintain consistent coding styles for multiple
developers working on the same project across various editors and IDEs.
The EditorConfig project consists of a file format for defining coding
styles and a collection of text editor plugins that enable editors to
read the file format and adhere to defined styles. EditorConfig files
are easily readable and they work nicely with version control systems."

This patch tries to configure the coding style seen so far. It's
indented to be updated, for example changing:

+insert_final_newline = false
+trim_trailing_whitespace = false

to
+insert_final_newline = true
+trim_trailing_whitespace = true

to remove whitespaces.

[1] https://editorconfig.org