Update ice_thermo_cpl.F90
By checking the subroutine "ice_surftemp" I found that t, the ice surface temperature, is limited to the freezing temperature of sea water (TFrezs, l.491).
If I understand the code correctly, this can be artificially wrong by up to 2 Kelvin, because ice is in principle allowed to reach 0°C at the surface (admittedly there is a small salinity content in the ice, but none in the snow). 

While this should help with getting lower ice thicknesses due to smaller atm-ice/snow heat flux (smaller temp gradient), it could also help with fixing the snow (which melts first).