IFS-interface: read precomputed weights matching mpi ranks
This makes fesom look for the precomputed interpolation weights
in a subdir with name 'dist_<mpi-ranks>' in the path given in
the namfesomcoup.in namelist as cdpathdist. (same as for mesh dist)
It is no longer required to change the namelist every time the
number of mpi ranks is changed.