fix(recom): correct idetn and idetc sms
Fix grazing contribution to idetn and idetc

Second detritus related parts are removed from the
computation of carbon (idetc) and nitrogen (idetn)
concentration of slow sinking detritus. Besides,
we should use grazEff instead of grazEff2
in the calculation of idetn. Because this is not related
to the number of the detritus class,
but to the zooplankton group / the source to this
detritus class.