Separate input and output
Separating the inputs from the outputs. This should enable us to have
more 'predictable' runs, in that running with the same inputs and the
same config/namelists should lead to the same results.
We can save time because we do not need to wait during backuping the
restart files (this is the main motivation in DestinE project).
Still this is done so that the old default behaviour is not changed.