feat(fesom): Part of atmospheric forcing as repeated year forcing
Change routines gen_surface_forcing.F90 and gen_modules_clock.F90
to allow selected number of years from atmospheric forcing such as
JRA55 to be used as repeated year forcing

We want to have repeat only part of the JRA55 forcing periodically (e.g repeat the period 1958-1967) through re-linking of the original forcing files.
1st problem: This will mess up the leap year cycle of the entire forcing, so
the model has to run best without leapyear although the calendar
system of the forcing remains gregorian (so with leapyear)
This can be solved by making the treatment of the time-axes
depending on the calendar of the forcing files, not alone from
the include_fleapyear flag.
2nd problem: In the case of a periodically repeated forcing period through
re-linking of the original files it happens that e.g. the year
1968 is linked to the file 1958 in that case time axis of the
1968 forcing file remains that of the original 1958 file. So
axis needs to be resetted properly.