Introduce cmake block for mesh partitioner to avoid build race condition
In this block set CMAKE_Fortran_MODULE_DIRECTORY to something different
from the 'normal' fesom module dir to not have a problem in a parallel
build where some module files are being written and read at the same
time (for modules that are shared between fesom main and mesh
partitioner). Would probably better to just link fesom as library to
mesh partitioner as well, but that currently implies other issues.