fix(recom): Correct silicate assimilation and diatom diagnostics 
We add __coccos directive to get the correct temperature
function in silicate assimilation.

In the coccos and Phaeocystis version, it should be dependent on the new temperature function, Temp_diatoms.
Use the original Arrhenius function in simpler phytoplankton version. Besides, light limitation diagnostics is corrected. It does affect only the diatoms related output.