fix(recom):  Return to saturation model for chlorophyll degradation
We commented out the linear model where degradation is directly
proportional to light intensity and the chlorophyll degradation
rate is capped at a maximum of 0.3. Reason for this is the unrealistic
air2sea co2 fluxes in the 2p3z2d setup.
The reactivated saturation model is more realistic. It accounts for light
saturation effects.
At low light, degradation increases roughly linearly with PAR
At high light, the degradation rate approaches an asymptotic maximum